исправлены ошибки при сохранении заказа в БД

+++ OrderItem.py
@@ -2,7 +2,7 @@
 
 class OrderItem:
 
-    def __init__(self, oil_name: str, volume: int, count: int, price: int):
+    def __init__(self, oil_name: str, volume: int, count: int, price: float):
         self.oil_name = oil_name
         self.volume = volume
         self.count = count
